Add turtlebot-demo jenkins job.
This adds a new job to the Jenkins config by adding turtlebot-demo as an
additional os_config. I looked for another place like `job_configs` to
do this since the turtlebot isn't a separate os per se but it looks like
`os_configs` is where this happens.

The job script will build a docker image with the
`INSTALL_TURTLEBOT2_DEMO_DEPS` variable set for turtlebot-demo jobs.

Then dockerfile has an added directive to install the demo dependencies
when that variable exists.

Still to do

- [ ] verify what other system package dependencies should be installed
- [ ] create a repos file (at https://raw.githubusercontent.com/ros2/turtlebot2_demo/master/turtlebot2_demo.repos?)
- [ ] run this in a non-dryrun capacity to test the job