Don't rely on hacking os_configs to create turtlebot job.
This pulls the turtlebot job out into its own section outside the
os jobs loop. It's possible down the line that this may be run within
the OS configs loop as another job. (ci_windows-turtlebot
ci_linux-turtlebot ...) but if that does happen we'll need to make it
easier for job config values to inject docker build params as right now
we use the same conditional block that switches the docker build
invocation based on os_name.

I think that refactoring is worth doing but it's something that can wait
to be paid down until we actually want this running on multiple
platforms.