Always set the --repo-file-url value.
Since the job creation script allows job-specific default repos urls, we
need to make sure those are honored by always passing the default repos
url.

As a temporary measure, this change is scoped to the turtlebot_demo job.
Once it tests fine there, I'll update this so it's not scoped to
turtlebot and add the same thing to the Windows bits.