Second pass at supplemental repos file implementation.
This factors curling and displaying the repos files into a function for
easy reuse.

The implementation as it is should allow us to fairly allow expansion
to N repos files. But I didn't add the functionality now for the reasons
in https://github.com/ros2/ci/pull/66#discussion_r113991571

The fact that repos files are iterated twice (once during download,
again during vcs import) is left intentionally unoptimized in order to
maintain the same structure in the build output as there would have been
with a single repos file.