Switch all % to ! in the windows batch script.
According to

https://social.technet.microsoft.com/Forums/scriptcenter/en-US/a009485a-046b-4dca-afd3-20555305e617/batch-variable-not-updating-as-expected?forum=ITCG

and empirical evidence, variable expansion is done at read time,
not execute time, by default in batch.  The way to make it work
at execute time is to setlocal enableDelayedExpansion (which we have),
and use ! for all variables (which we don't have).  Switch all
of these to ! so that we get expansion at execute time.  I can't
say I understand why this worked before; it should have always
had this problem.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>