Use triple dashes at the end of the build args. (#86)
* Use triple dashes at the end of the build args.

This ensures that the -- makes it onto the command-line,
and should fix jobs that sometimes gets the arguments
embedded into the directory names.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Switch all % to ! in the windows batch script.

According to

https://social.technet.microsoft.com/Forums/scriptcenter/en-US/a009485a-046b-4dca-afd3-20555305e617/batch-variable-not-updating-as-expected?forum=ITCG

and empirical evidence, variable expansion is done at read time,
not execute time, by default in batch.  The way to make it work
at execute time is to setlocal enableDelayedExpansion (which we have),
and use ! for all variables (which we don't have).  Switch all
of these to ! so that we get expansion at execute time.  I can't
say I understand why this worked before; it should have always
had this problem.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Escape more correctly.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Fix windows brokeness.

Regular for loops always split on space and equals signs, and
there is nothing you can do about it.  Switch to /f instead
and do the loop ourselves, which is more ugly but actually
works.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>

* Reinstate ARCH= variable.

We don't know for sure it is not needed anymore, so just
leave it alone for now.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>