Cleanup the loop in create_jenkins_job.py that creates jobs.
In the previous version of the loop, the job_data dict was
shared amongst all jobs.  This means that later jobs could
unintentionally inherit properties of earlier jobs.  Instead,
introduce a small function to do the actual job creation,
which ensures we get a fresh dict every time.  That is,
we start off with the standard dict in data, update it with
OS-specific stuff from the os_configs dict, and then update
it with job-specific data passed into the create_job()
function.  This moves duplicated logic out of the loop
and ensures that jobs don't inherit properties from other
jobs.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>