Set default CI_ISOLATED parameter to false on Windows.
Jobs using isolation are starting to fail on Windows as we're coming up
against the maximum environment variable and input line lengths.

Since other platforms are still building with isolation we don't lose
too much coverage.

Adds a template variable for the default setting of the CI_ISOLATED
parameter and keeps it set to true by default (previously the default
was hard-coded) and sets it to false for Windows only. It's still
possible to request a build with isolation by changing the build
parameter when starting a new build.