Create a test job matching each ci job.
Adds configuration updates for the test_ci_* jobs that we sometimes use
for the build farmer to test platform changes or job configuration
changes while leaving the rest of the CI cluster intact for regular use.

So far I have intentionally left out:
* test_packaging jobs: I think the existing ci_packaging jobs serve this
purpose. We could rename the ci_packaging jobs to test_packaging (or
leave them as is) if that's the case or make additional changes to
create both ci_packaging and test_packaging jobs.
* Changing the default label expression for test_ jobs: test_ jobs are
identical to ci_ jobs by default. We could automatically add a
test_$PLATFORM label for the test jobs but it might be just as easy to
leave it alone as the build farmer will need to apply the label to the
nodes under test and it may be that we want to test some long-running
changes with a persistent label (i.e. bionic or high-sierra).