Create a test job matching each ci job. (#161)
* Create a test job matching each ci job.

Adds configuration updates for the test_ci_* jobs that we sometimes use
for the build farmer to test platform changes or job configuration
changes while leaving the rest of the CI cluster intact for regular use.

Intentionally left unchanged are:

* ci_packaging jobs: I think the existing ci_packaging jobs serve as testbeds for packaging changes. We could rename the ci_packaging jobs to test_packaging or create additional test_packaging jobs in the future.
* The default label expression for test_ jobs: test_ jobs are
identical to ci_ jobs by default. We could automatically add a
test_$PLATFORM label for the test jobs but it might be just as easy to
leave it alone as the build farmer will need to apply the label to the
nodes under test and it may be that we want to test some long-running
changes with a persistent label (i.e. bionic or high-sierra).

Thanks Mikael for review and typo catching.