Modify ASAN/TSAN nightly jobs configuration
This change modifies the ASAN and TSAN jobs to
only run until rcutils.

I chose rcutils because it is already
green for TSAN and will be green soon for ASAN.

Once both nightly jobs are green, we will have a way
to track ASAN/TSAN regressions.

As the team continues to "clean-up" new packages
higher in the stack, we will continue editing the
nightly jobs to expand progressively the scope of those
jobs to include "newly green" packages.

Signed-off-by: Thomas Moulard <tmoulard@amazon.com>