Set default retention to 1000 days, 3000 builds. (#283)
* Set default retention to 1000 days, 3000 builds.

We set this retention policy on ci_* jobs in #278.  This makes that the
default for all generated jenkins jobs (keeping the 180 days, 100 builds
limit on ci_packaging jobs) which will propagate the effect to nightlies
and packaging jobs.

* Set 370 day retention on packaging jobs.

Per Dirk's point that long term the math doesn't add up to keep 1000
days of packaging jobs.