Attempt to find two different possible RTI setenv scripts.
The script name is dependent on the platform and compiler version. I
think the "mojave" version is actually the High Sierra version of
Darwin. I received a warning that the next version of macOS (Catalina)
would not support this binary but Mojave appears to so we're good there
for now.