Use version rather than release id as upstream Docker image tag.
The 1909 image doesn't provide sufficient stability for our instances.
Newly deployed agents with a newer version of Windows 1909 may not be
compatible with the pinned image and the 1909 release id may move
forward to become incompatible with an existing AMI.

In testing on both new and old agents fetching the release version and
using it as the image tag resolved the issues installing python inside
the container.