Pin flake8-blind-except to 0.1.1
A new release of flake8-blind-except is causing a fair number of
packages to fail their flake8 tests, where they were passing when
flake8-blind-except 0.1.1 was used.

This has started a conversation about pristine testing environments and
the influence of these flake8 plugins that aren't actually declared as
dependencies. This is the lowest-effort way to suppress the issue until
that discussion has concluded.