Return to using qt online installer by default.
This change in tandem with refactored qt5 recipes will allow us to use
either the qt5 online installer with the latest version of qt5 on
Windows or the 5.12 offline installer. This is designed to be used
whenever the Qt5 online installer is unavailable. Since the offline
installer stopped being made available for open source builds after
5.12 it is not recommended unless the online installer is unusable.