Update xunit configuration tag name for report types to tools.
During the upgrade from Jenkins 2.263.3 to 2.277.3 jobs began to fail
due to improper XML schema in the xunit configuration.

Despite the fact that xunit itself has not changed this appears to be
required for function.

This change has not been tested on Jenkins <2.277.3 and so a Jenkins
upgrade may be required to allow this to work.