Enable RTI Connext on RHEL 8
It seems that the binaries supplied by RTI are intended to be somewhat
portable among Linux platforms, and there is enough similarity to
utilize Connext on RHEL 8 in the same way we've used it in Ubuntu since
Bionic.

This cleanup of the RHEL RMW configuration values has the additional
benefit of adding rmw_fastrtps_dynamic_cpp to the packaging job, which
aligns it better with the Ubuntu archives.

Signed-off-by: Scott K Logan <logans@cottsay.net>