For RHEL, install specific colcon packages
When CI uses a venv, these specific colcon packages are installed. When
installing colcon-common-extensions, other colcon packages like
colcon-notification get installed and are not installed in the venv,
leading to a mixed environment of system packages and pip-installed
colcon packages.

Specifically installing the colcon packages that CI uses in the venv
should help us maintain a homogeneous and reproducible environment.

Signed-off-by: Scott K Logan <logans@cottsay.net>