Remove --do-venv from CI and packaging jobs.
Versions of virtualenv past 16.7.9 don't respect the
install_scripts settings in setup.cfg.  The practical effect
of this is that colcon, and the rest of our tools, are looking
in the wrong place for scripts.

Unfortunately, our pin of virtualenv 16.7.9 does not work when
we are using Python 3.10 or greater.

For these reasons, remove using a virtualenv in the linux jobs
altogether.  Since we are already in a docker container, we
don't really need another level of isolation anyway.

Signed-off-by: Chris Lalancette <clalancette@openrobotics.org>