Use container image snapshots running Python 3.9 on Jammy.
This is a barebones commit which pulls and runs Ubuntu Jammy containers
rather than building an image during the CI job.

In general this is not a preferred or recommended approach but something
that is being implemented to make sure that CI stays available while we
work on the changes required to bring CI online for Python 3.10.

Notably absent in this first commit is any kind of support for running
non-Jammy builds.