Set workdir and entrypoint at runtime.
When trying to operate CI off of a pre-existing snapshot one of
the things holding us back is that the docker image is dynamically
rebuilt with references to the current job name in paths. Rather than
normalize the path across jobs or fight this, let's use arguments to
docker-run to set these things for each container run.

Because we can't change the snapshots this also makes it easier to tweak
the entry_point script used by them.