Remove spurious directory change.
This `cd @workdir` is redundant with the WORKDIR declaration with which
it was added and with the use of `--workdir` in the current patches.

By eliminating this we remove a source of job-specific content in the
scripts and Dockerfile contents for CI.

By removing this we're also committing ourselves to being extra vigilant
that behavior inside the entry_point.sh script does not change the
directory context without restoring it. If there was concern about that
then an alternative would be to add

	BACKUPPWD=`pwd`
	# ... script contents
	cd $BACKUPPWD

instead of this change.