Use container image snapshots running Python 3.9 on Jammy. (#640)
* Use container image snapshots running Python 3.9 on Jammy.

This is a barebones commit which pulls and runs Ubuntu Jammy containers
rather than building an image during the CI job.

In general this is not a preferred or recommended approach but something
that is being implemented to make sure that CI stays available while we
work on the changes required to bring CI online for Python 3.10.

Notably absent in this first commit is any kind of support for running
non-Jammy builds.

* Only use the jammy snapshot container for jammy.

These changes would completely clobber a job running focal otherwise.

* Set workdir and entrypoint at runtime.

When trying to operate CI off of a pre-existing snapshot one of
the things holding us back is that the docker image is dynamically
rebuilt with references to the current job name in paths. Rather than
normalize the path across jobs or fight this, let's use arguments to
docker-run to set these things for each container run.

Because we can't change the snapshots this also makes it easier to tweak
the entry_point script used by them.

* Make entry_point.sh executable in-repo.

Instead of changing the mode once in the dockerfile.

* Add back the default entrypoint.

Setting entrypoint via the command line purges the default cmd value
which we want to preserve. It ought to be enough to bind-mount the
entry_point.sh script.

* Fix typo in entry_point.sh bind mount.

* Quote variables to protect against undefined variables.

Co-authored-by: Scott K Logan <logans@cottsay.net>

Co-authored-by: Scott K Logan <logans@cottsay.net>