Set workdir at runtime and remove sed scripts from Dockerfiles and entry_point.sh. (#641)
* Set container working directory and bind-mount entry_point script.

This allows the working directory and entry point script to be dynamically defined which decreases the differences between container base images between jobs.

The entry_point.sh script is mounted read-only inside the container even
though I don't think that is critical.

* Make the entry_point.sh script exectuable in the repository.

We were previously changing the permissions once copied into the
container but now that it's being bind-mounted it should be executable
in the repository.

* Remove Dockerfile directives which are set at runtime.

* Remove spurious directory change.

This `cd @workdir` is redundant with the WORKDIR declaration with which
it was added and with the use of `--workdir` in the current patches.

By eliminating this we remove a source of job-specific content in the
scripts and Dockerfile contents for CI.

By removing this we're also committing ourselves to being extra vigilant
that behavior inside the entry_point.sh script does not change the
directory context without restoring it. If there was concern about that
then an alternative would be to add

	BACKUPPWD=`pwd`
	# ... script contents
	cd $BACKUPPWD

instead of this change.

* Remove basepath sed scripts.

Hard-coded references to the working directory have been replaced in
favor of runtime arguments to docker-run.

* Copy entry_point.sh rather than bind mount it.

Now that the script has no job-specific components it makes sense again
to just copy it into the container.