Use system versions of setuptools/pip unless there is none available. (#643)
* Install setuptools and pip from apt.

Installing these packages from pip results in CI volatility.  setuptools
60.10.0 appears to interfere with the install paths, unconditionally
inserting `local` in the path so we're left with `$PREFIX/local/bin`
rather than `$PREFIX/bin` as expected.

pip is somewhat of a coattail rider on this commit but is being sourced
from apt for similar reasons.

I've omitted the installation of virtualenv entirely because of parallel
patches currently in flight. That may be revisited before this is ready
for merge.

Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>

* Only update setuptools and pip from pypy if setuptools is not yet installed.

I think this may even be dead code with MacOS disabled:
* Ubuntu installs setuptools and pip from apt during container build.
* RHEL installs setuptools and pip from dnf during container build.
* Windows has setuptools and pip installed via chef during container
  build.

So it may actually make sense to stop this altogether assuming that
using system setuptools works for Foxy, Galactic, and Rolling on Focal.

Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>

* Remove dead code.

As previously discussed, setuptools is now available on all our
supported platforms so there's no need to pull a newer version from
pypi.

Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>