Opt-out of Phased Updates in apt.
Docker containers are built using a succession of images, each of of
which is created by running an intermediate container and saving the
result.

These individual containers can end up with different information that
puts them in different "phases" of the apt phased update rollout.

When combined with a phased update for a set of binary packages that all
rely on matching versions, it can create an unsatisfiable condition
where an earlier binary from the source package was installed in the
"ahead" phase and a later package is installed "behind".

Phased updates only became supported in apt 2.1.16 released in Ubuntu
21.04 so this was not an issue for us prior to Ubuntu Jammy.