Switch to always opting into phased updates. (#709)
When initially installing a package, phased update percentage is not
considered at all. It's only used when considering package upgrades.

As a result, opting out of phased updates can create a situation where a
package from the updates repository is installed during the container
build in an earlier phase, but a later installation may try to install
dependencies in sibling packages which are not available at the required
version due to being behind the phased update percentage wall.