Revamp --ignore-rmw handling. (#715)
* Revamp --ignore-rmw handling.

The job templates are currently always setting
'--ignore-rmw rmw_connext_dynamic_cpp', and then adding
to that list if other RMWs should be ignored.
But it is not necessarily the case that we will always
have an RMW to ignore.  Therefore, we switch to passing
--ignore-rmw for *each* of the ones that we want to ignore
(because there may not be any). To handle that, change the
argparse handling of --ignore-rmw to use the 'extend' action,
and then modify the entrypoint.sh to deal with the fact that
there might be multiple --ignore-rmw flags, any of which could
contain 'connext'.

* Also handle ignoring rmw_connext_cpp in the entrypoint.

I'll note that the existing code is actually kind of buggy.
That is, unless you specify --ignore-rmw on *both* rmw_connext_cpp
and rmw_connextdds, Connext will be installed.  This commit
actually retains that behavior, but that will be fixed by
upcoming PRs.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>