Remove support for Connext 5.3.1.
There are no current distributions that currently require it.

One thing in this commit that warrants some explanation is
the complete removal of the sourcing of Connext in the
linux_batch/__init__.py and windows_batch/__init__.py files.
They are completely removed because with Connext 6, they
are no longer used.  On Linux it is sourced from the entry_point.sh,
and on Windows I'm not entirely sure where it is sourced
from.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>