Drop cppcheck test exclusion on RHEL
Previously, the version of cppcheck distributed in RHEL resulted in many
(unactionable) linter failures only present on RHEL. Given the limited
utility of running the linters on RHEL to begin with, we decided to
simply skip the tests.

Changes either to ament_cppcheck or the version of cppcheck distributed
in RHEL now appear to allow these tests to run successfully, and the
cost of maintaining this special case now outweigh the benefit.

Reverts part of c0973c628ec15.