Always create constraints file.
That is, even if we aren't going to install a package
(since it was already installed), make sure to add the
the number to the constraints file so that other installations
won't upgrade it.

While we are here, print out the constraints in use for
debugging further down the line.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>