Redo how we install pip packaging. (#764)
* Redo how we install pip packaging.

The current way we install packages in __main__ is really
a mess.  Some of the packages have pinned versions, some do
not.  Some of the packages are installed only if they are
not installed by the underlying OS, some are not.

This commit partially cleans this up.  In particular, for
all pip packages that are *not* colcon-related, we have a tuple
which contains the package name, the import name, and the (optional)
version we want installed.  For every pip package name, we see
if we can import it (using the import name); if we can,
then we don't need to install it and we skip it.  If we
*do* need to install it, then we add it to the list.  Even if we are
not going to install the package, make sure add any constraints
so that the installation of other packages will respect the constraint.

This should mean that this code works on all of Ubuntu, RHEL,
and Windows.  On Ubuntu and RHEL, almost all packages are
installed via the Dockerfiles (the one exception being colcon).
For Windows, almost nothing is installed by the chef scripts,
so most things will be installed here.

A further cleanup here would to also install colcon from
packages rather than pip, but this is large enough as it is.
If this is acceptable, we can do that in a followup.

Along with some other fixes, this change allows us to actually
launch an Ubuntu 24.04 job (it still won't pass tests, but that
is due to code-related reasons which we are fixing separately).

Signed-off-by: Chris Lalancette <clalancette@gmail.com>
Co-authored-by: Scott K Logan <logans@cottsay.net>