Fix a couple more problems found while testing CI.
1.  Make sure to install the connext debians on both jammy
and noble.  This doesn't make a huge difference on the CI
jobs (where we'll install Connext via scripts later anyway),
but matters for the packaging jobs.
2.  Revamp how we are installing the LTTng-related packages
on RHEL.  The logic here is complex enough that it warrants
splitting it out to a separate block.  To wit, on Humble
we install nothing (ros2_tracing is not installed by default),
on Iron we install the dependencies plus python3-lttng (we are
using the Python bindings as provided by the platform), and on
Jazzy and newer we install the lttng-tools-devel package so we
can build our own Python bindings.  This also makes the RHEL
logic match the Ubuntu logic.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>