Fix a couple more problems found while testing CI. (#773)
* Fix a couple more problems found while testing CI.

1.  Make sure to install the connext debians on both jammy
and noble.  This doesn't make a huge difference on the CI
jobs (where we'll install Connext via scripts later anyway),
but matters for the packaging jobs.
2. Slightly revamp how we install the LTTng packages for RHEL.
In particular, on RHEL-8, don't install anything since the packages
are too old there.  On RHEL-9, install the union of all packages
necessary for Iron, Jazzy, and Rolling, and the individual code
will just use the correct package as needed.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>