Install RHEL-8 pip packages in Dockerfile-RHEL.
It turns out that on RHEL-8, we need to install
a handful of packages from pip.  Doing this means
that the pip installation within __main__.py will
do exactly nothing, since all dependencies are
already installed.

This actually fixes two bugs:
1.  There is currently a bug where trying to run
'pip install foo' on RHEL-8 hits a permission denied
error.  This change avoids it by running that pip
install when we are effectively root in the container.
2.  Installing a too new version of pycodestyle doesn't
work with the pinned pytest that we are using.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>