Multi Visual Studio version support  (#786)
* Add in the ability to switch MSVC versions.

In particular, we add in the ability to switch between
MSVC 2019 and 2022.  To get there requires a lot of
explanation.

First of all, we need to modify the Chef recipes over
at https://github.com/ros-infrastructure/ros2-cookbooks
how to install alternate versions of MSVC and how to
setup the environment for each.  That is done by the
companion PR to this one.

In this PR, we need to tell the Chef recipes which version
to install, and then also kick off the build using the
appropriate environment.

In order to tell the Chef recipes to install, we use a
configuration variable called "vs_release" in the
install_ros_<rosdistro>.json file per ROS distribution.
The value we set for the "vs_release" is always "@vs_version";
that makes substitution later easier.  Then in the job
configuration, we use a string replace to replace the
@vs_version with the actual version of MSVC we want Chef to
install (I attempted to use a regex here, but I ran into
nasty quoting issues between cmd, powershell, and Docker,
and finally gave up).

The other part of this change is to kick off the build using
the appropriate environment.  It turns out that we have been
double-setting up the environment via vcvarsall.bat on Windows
for a long time; once in the Dockerfile, and once in the
ros2_batch_job batch script which setup the environment.  However,
the latter version was buggy, and actually pointed to the incorrect path.

In theory, we could fix either vsvarsall.bat invocation.  In
practice, I ran into (more) nasty issues with being able to
do environmental substitutions on Windows during a Docker
CMD.  To sidestep this completely, we remove the setup of the
environment there, and fix the pathing issues in ros2_batch_job.
This seems to make everything work.

* A couple more fixes.

First, set 2022 *after* 2019, so that 2019 remains the
default for now.

Second, make sure to use @@ in the job_templates for the
deploy, otherwise create_jenkins_jobs.py barfs when trying
to do a substitution via em.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>