Pin the version of pixi.
The main reason to do this is to make sure that we aren't
subject to man-in-the-middle attacks, where we are served
a malicious version of pixi.  Instead, we have a hard-coded
sha256 of the binary, and if it fails to match the build
will fail.  This also means that if we want to update the
version of pixi, we'll have to manually update the hash.

Signed-off-by: Chris Lalancette <clalancette@gmail.com>