Fetch pixi.toml directly from a git repository clone.
Using an in-container command to fetch this file prevents docker from
invalidating cached copies of the fetch when the remote content changes.

Cloning the git repository may result in overly aggressive cache
busting, but using ADD with the githubusercontent url directly will also
have problems during iteration as that url is also cached for a period
of time and does not update immediately when repository contents change.