Restore daily invalidation of Windows Dockerfile. (#807)
This was removed in #802 when removing the chef configuration but we do
this on Linux as well.

It makes sense to do this before configuring the pixi workspace if not
also doing before pixi installation.

In the chef configuration, due to the amount of time spent in
installation we did one chef provision before invalidation and then one
after it in the hope that unchanged things would stay unchanged and the
catchup run would not take as much time each day.

On linux we invalidate after initial configuration but before
dependency installation so we could move this up in the Windows
container if we want.

Also I renamed the sigil string from `todays_date` to `today_str` so
that the sigil we use on Linux and Windows match.